# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: سند – منصة الخدمات والمهنيين في اليمن
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List requests (for current user or provider)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, accepted, rejected, in_progress, completed, cancelled, null]
        - name: role
          in: query
          schema:
            type: ["string", "null"]
            enum: [client, provider, null]
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a new service request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get request details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"

  # ─── REVIEWS ─────────────────────────────────────────────────────────────────
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /providers/{id}/reviews:
    get:
      operationId: getProviderReviews
      tags: [reviews]
      summary: Get reviews for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # ─── FAVORITES ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ServiceRequest:
      type: object
      required: [id, clientId, providerId, status, serviceType, description, city, district, scheduledAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientAvatarUrl:
          type: ["string", "null"]
        providerId:
          type: integer
        providerName:
          type: string
        providerAvatarUrl:
          type: ["string", "null"]
        providerCategoryName:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, in_progress, completed, cancelled]
        serviceType:
          type: string
        description:
          type: string
        city:
          type: string
        district:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        scheduledAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        isImmediate:
          type: boolean
        createdAt:
          type: string
    ServiceRequestInput:
      type: object
      required: [providerId, serviceType, description, city, district, isImmediate]
      properties:
        providerId:
          type: integer
        serviceType:
          type: string
        description:
          type: string
        city:
          type: string
        district:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        scheduledAt:
          type: ["string", "null"]
        isImmediate:
          type: boolean
    ServiceRequestUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
          enum: [accepted, rejected, in_progress, completed, cancelled, null]

    # ── REVIEW ──────────────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, clientId, providerId, rating, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientAvatarUrl:
          type: ["string", "null"]
        providerId:
          type: integer
        requestId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [providerId, rating]
      properties:
        providerId:
          type: integer
        requestId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]

    # ── FAVORITES ──────────────────────────────────────────────────────────────
...